backend/requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.35.0
sqlalchemy[asyncio]==2.0.41
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.16.4
pydantic==2.11.7
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
loguru==0.7.3
scapy==2.6.1
redis==5.2.1
email-validator==2.2.0
orjson==3.10.18
pytest==8.4.1
pytest-asyncio==1.1.0
httpx==0.28.1
ruff==0.12.7
mypy==1.17.0
